[fix] #128 - 상품 개수 불일치 문제 해결(isVisible 적용 안되어있었음)

+++ core-domain/src/main/java/com/napzak/domain/store/vo/StoreStatus.java
@@ -4,15 +4,15 @@ public record StoreStatus(
 	Long userId,
 	String storePhoto,
 	String nickname,
-	Long totalSellCount,
-	Long totalBuyCount
+	int totalSellCount,
+	int totalBuyCount
 ) {
 	public static StoreStatus from(
 		Long userId,
 		String storePhoto,
 		String nickname,
-		Long totalSellCount,
-		Long totalBuyCount
+		int totalSellCount,
+		int totalBuyCount
 	) {
 		return new StoreStatus(userId, storePhoto, nickname, totalSellCount, totalBuyCount);
 	}
